+++ client/src/Events.js
@@ -1,0 +1,16 @@
+export default {
+  _cbs: [],
+  notify (eventName) {
+    this._cbs.map(cb => {
+      if (cb.eventName === eventName && typeof cb.callback === 'function') {
+        cb.callback()
+      }
+    })
+  },
+  on (eventName, cb) {
+    this._cbs.push({
+      eventName: eventName,
+      callback: cb
+    })
+  }
+}