+++ client/src/index.js
@@ -10,14 +10,25 @@ import verifyEmail from './verifyEmail'
 import loginWithFacebook from './oauth/loginWithFacebook'
 import loginWithGoogle from './oauth/loginWithGoogle'
 import loginWithLinkedIn from './oauth/loginWithLinkedIn'
+import loginWithVK from './oauth/loginWithVK'
 import userId from './userId'
-import {onTokenChange, getLoginToken, setTokenStore} from './store'
+import {
+  initWithClient,
+  setTokenStore,
+  getLoginToken,
+  onLogin,
+  onLoginFailure,
+  onLogout,
+  loggingIn,
+  TOKEN_EXPIRES_KEY,
+  TOKEN_KEY,
+  USER_ID_KEY
+} from './store'
 
 export {
   changePassword,
   createUser,
   forgotPassword,
-  getLoginToken,
   hashPassword,
   loginWithPassword,
   logout,
@@ -27,7 +38,42 @@ export {
   loginWithFacebook,
   loginWithGoogle,
   loginWithLinkedIn,
-  onTokenChange,
+  loginWithVK,
+  userId,
+  initWithClient,
   setTokenStore,
-  userId
+  getLoginToken,
+  onLogin,
+  onLoginFailure,
+  onLogout,
+  loggingIn,
+  TOKEN_EXPIRES_KEY,
+  TOKEN_KEY,
+  USER_ID_KEY
+}
+
+export default {
+  changePassword,
+  createUser,
+  forgotPassword,
+  loginWithPassword,
+  logout,
+  resendVerificationEmail,
+  resetPassword,
+  verifyEmail,
+  loginWithFacebook,
+  loginWithGoogle,
+  loginWithLinkedIn,
+  loginWithVK,
+  userId,
+  initWithClient,
+  setTokenStore,
+  getLoginToken,
+  onLogin,
+  onLoginFailure,
+  onLogout,
+  loggingIn,
+  TOKEN_EXPIRES_KEY,
+  TOKEN_KEY,
+  USER_ID_KEY
 }