+++ client/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+### vNEXT
+
+- Add `initWithClient(apolloClientInstance)` function to remove need for passing apollo with every function call
+- Calling `userId()` is now synchronous and more performant since it's cached in memory
+- Added the following hooks/callbacks: `onLogin`, `onLoginFailure`, `onLogout` so the API is similar to Meteor Accounts
+- Added `loggingIn` function, similar to Meteor Accounts
+- Added support for persisting login automatically. As soon as you call `initWithClient` in your app, it will check for a stored token and attempt to login, calling the `onLogin` callback if it succeeds.
+### v2.1.0
+- Bump dependencies
+
 ### v2.0.2
 
 - Improve logout function.