+++ meteor-server/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+### vNEXT
+
+- Added `loginWithToken` mutation
+
+### v3.2.3
+
+- Login with VK.
+
 ### v3.1.0
 
 - Use http instead of meteor to make Facebook login.